scrapers/requirements_lambda.txt
# Lambda-specific requirements
# Note: These should be packaged as Lambda layers or included in deployment package

# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dateutil>=2.8.0
certifi>=2022.12.7
urllib3>=1.26.0

# PostgreSQL (if writing to RDS)
psycopg2-binary>=2.9.0


